requirements.txt

Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.6
redis==5.0.1
Pillow==10.2.0
django-extensions==3.2.3
